+++ com.unity.mobile.notifications/Editor/UnityNotificationEditorManager.cs
@@ -33,6 +33,17 @@ namespace Unity.Notifications
         Alert = 1 << 2,
         All = ~0,
     }
+    
+    [Flags]
+    internal enum AuthorizationOptionEditor
+    {
+        Badge = 1 << 0,
+        Sound = 1 << 1,
+        Alert = 1 << 2,
+        CarPlay = (1 << 3),
+        All = ~0,
+    }
+
 
     [System.Serializable]
     internal class DrawableResourceData
@@ -177,14 +188,26 @@ namespace Unity.Notifications
             }
             set
             {
+                string strValue;
+
+                if (value is Enum)
+                {
+                    strValue = ((int) value).ToString();
+                }
+                else
+                {
+                    strValue = value.ToString();
+                }
+                
+                
                 var index = keys.IndexOf(key);
                 if (index == -1)
                 {
                     keys.Add(key);
-                    values.Add(value.ToString());
+                    values.Add(strValue);
                 }
                 else
-                    values[index] = value.ToString();
+                    values[index] = strValue.ToString();
             }
         }
     }
@@ -209,13 +232,11 @@ namespace Unity.Notifications
 
         private void SaveSetting(NotificationEditorSetting setting, NotificationEditorSettingsCollection values)
         {
-            if (values == null )
-                values = new NotificationEditorSettingsCollection();
-
             if (!values.Contains(setting.key) || values[setting.key].ToString() != setting.val.ToString())
             {
+                
                 values[setting.key] = setting.val;
-                SerializeData();
+                EditorUtility.SetDirty(this);
             }
         }
 
@@ -239,9 +260,6 @@ namespace Unity.Notifications
         
         public T GetAndroidNotificationEditorSettingsValue<T>(string key, T defaultValue)
         {
-            if (AndroidNotificationEditorSettingsValues == null)
-                AndroidNotificationEditorSettingsValues = new NotificationEditorSettingsCollection();
-
             try
             {
                 var val = AndroidNotificationEditorSettingsValues[key];
@@ -250,7 +268,7 @@ namespace Unity.Notifications
             }
             catch (InvalidCastException)
             {
-                AndroidNotificationEditorSettingsValues = new NotificationEditorSettingsCollection();
+                //Just return default value if it's a new setting that was not yet serialized.
             }
 
             AndroidNotificationEditorSettingsValues[key] = defaultValue;
@@ -261,9 +279,6 @@ namespace Unity.Notifications
         public T GetiOSNotificationEditorSettingsValue<T>(string key, T defaultValue)
         {
 
-            if (iOSNotificationEditorSettingsValues == null)
-                iOSNotificationEditorSettingsValues = new NotificationEditorSettingsCollection();
-
             try
             {
                 var val = iOSNotificationEditorSettingsValues[key];
@@ -272,7 +287,8 @@ namespace Unity.Notifications
             }
             catch (InvalidCastException)
             {
-                iOSNotificationEditorSettingsValues = new NotificationEditorSettingsCollection();
+                Debug.LogWarning("Failed loading : " + key + " for type:" + defaultValue.GetType() + "Expe cted : " + iOSNotificationEditorSettingsValues[key].GetType());
+                //Just return default value if it's a new setting that was not yet serialized.
             }
 
             iOSNotificationEditorSettingsValues[key] = defaultValue;
@@ -355,6 +371,17 @@ namespace Unity.Notifications
                 AssetDatabase.CreateAsset(notificationEditorManager, assetRelPath);
                 AssetDatabase.SaveAssets();
             }
+            
+            if (notificationEditorManager.iOSNotificationEditorSettingsValues == null)
+            {
+                notificationEditorManager.iOSNotificationEditorSettingsValues = new NotificationEditorSettingsCollection();
+            }
+
+            if (notificationEditorManager.AndroidNotificationEditorSettingsValues == null)
+            {
+                notificationEditorManager.AndroidNotificationEditorSettingsValues =
+                    new NotificationEditorSettingsCollection();
+            }
 
             var iosSettings = new List<NotificationEditorSetting>()
             {
@@ -371,9 +398,9 @@ namespace Unity.Notifications
                             "UnityNotificationDefaultAuthorizationOptions",
                             "Default Notification Authorization Options",
                             "Configure the notification interaction types your app will include in the authorisation request  if  “Request Authorisation on App Launch” is enabled. Alternatively you can specify them when creating a `AuthorizationRequest` from a script.",
-                            notificationEditorManager.GetiOSNotificationEditorSettingsValue<PresentationOption>(
+                            notificationEditorManager.GetiOSNotificationEditorSettingsValue<AuthorizationOption>(
                                 "UnityNotificationDefaultAuthorizationOptions",
-                                (PresentationOption) PresentationOptionEditor.All)
+                                (AuthorizationOption) AuthorizationOption.Alert | AuthorizationOption.Badge | AuthorizationOption.Sound)
                         ),
 
 
@@ -421,6 +448,7 @@ namespace Unity.Notifications
                     false),
             };
 
+            
             if (notificationEditorManager.iOSNotificationEditorSettings == null ||
                 notificationEditorManager.iOSNotificationEditorSettings.Count != iosSettings.Count)
             {